refactor: CommunityUseCaseをusecase/communityに移動
